Stores actions
